+++ src/Entities/Buildings/CraftBuilding.h
@@ -12,7 +12,7 @@ class CraftBuilding : public Building {
     void stopRecipe();
 
 public:
-    CraftBuilding(int id, std::pair<double, double> pos, ptr<Player> owner, unsigned int hp, double radius = baseBuildingRadius,
+    CraftBuilding(int id, std::pair<double, double> pos, ptr<Player> owner, double hp, double radius = baseBuildingRadius,
                   ptr<Building> parent = ptr<Building>())
         : Building(id, pos, owner, hp, radius, parent),
           Entity(pos, owner, hp, radius),
